Fix null response in GetFamilyGroupInformation endpoint
