+++ FamilyVaultServer/Services/PrivMx/Models/Result/PrivMxGetContextInfo.cs
@@ -1,0 +1,10 @@
+using System.Text.Json.Serialization;
+
+namespace FamilyVaultServer.Services.PrivMx.Models.Result
+{
+    public class PrivMxGetContextResult : PrivMxResponseResult
+    {
+        [JsonPropertyName("context")]
+        public PrivMxGetContext_ContextInfo ContextInfo { get; set; }
+    }
+}
